+++ ljepotaservis/ljepotaservis.Web/ClientApp/src/components/userSide/Store.js
@@ -1,7 +1,20 @@
 import React, { Component } from "react";
+import { connect } from "react-redux";
+import RoleNavbar from "../navbar";
 
-export default class Store extends Component {
-  render() {
-    return <div>id: {this.props.match.params.id}</div>;
-  }
-}
+const Store = props => {
+  return (
+    <div>
+      <RoleNavbar role={props.role} />
+      <div>id: {props.match.params.id}</div>;
+    </div>
+  );
+};
+
+const mapStateToProps = state => ({
+  role: state.authentication.hasOwnProperty("user")
+    ? state.authentication.user.role
+    : "Guest"
+});
+
+export default connect(mapStateToProps)(Store);
