+++ ljepotaservis/ljepotaservis.Web/ClientApp/src/utils/handleResponse.js
@@ -1,0 +1,18 @@
+import { userActions } from "../store//actions/userActions";
+
+export const handleResponse = response => {
+  return response.text().then(text => {
+    const data = text && JSON.parse(text);
+    if (!response.ok) {
+      if ([401, 403].indexOf(response.status) !== -1) {
+        // auto logout if 401 Unauthorized or 403 Forbidden response returned from api
+        userActions.logout();
+      }
+
+      const error = (data && data.message) || response.statusText;
+      return Promise.reject(error);
+    }
+
+    return data;
+  });
+};
